refactor(keymap): separate keycodes.ts data and logic (2152 lines)
Split keycodes.ts into data definitions and utility functions:
- keycodes-types.ts: KeycodeOptions, CustomKeycodeDefinition, KeyboardKeycodeContext
- keycodes-utils.ts: all logic functions (resolve, serialize, findKeycode, etc.)
- keycodes.ts: Keycode class + data arrays only, re-exports for backward compat

Data arrays remain in keycodes.ts (no line limit per project rules).
All existing import paths preserved via re-exports.
